requirements.txt
pint
pandas
seaborn
numpy
IPython
ipykernel
tabulate
python-docx
matplotlib
openpyxl
devtools
types-tabulate